backend/requirements/dev.txt
-r base.txt

django-debug-toolbar==3.6.0
mysqlclient==2.1.1
